Added Games log and method control
Added Games tab in tensorboard to see number of games during training. Also added all gigar, gicar, gidb and gib methods in minigrid.py. Method is choosed by passing to seed method of MinigridEnv in training script variable strat_method as a string of one of the others. Keep an eye if you correctly inserted the name of a method.